requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.20.42
botocore==1.23.42
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.11
django-ckeditor==6.3.2
django-contrib-postgres==0.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-rest-knox==4.1.0
django-tenants==3.4.2
djangorestframework==3.13.1
dynaconf==3.1.7
environ==1.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-secret-manager==2.10.0
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.46.0
grpcio-status==1.46.0
gunicorn==20.1.0
httplib2==0.20.4
hvac==0.11.2
idna==3.3
importlib-metadata==4.10.1
Jinja2==3.0.3
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
paramiko==2.9.2
Pillow==9.0.0
proto-plus==1.20.3
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
s3transfer==0.5.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
validators==0.18.2
whitenoise==5.3.0
zipp==3.7.0
